Split examples into separate files (#1456)
